+++ examples/server/src/resolvers/index.ts
@@ -2,12 +2,12 @@ import { ResolverMap } from "../generated/schema";
 import { MyContext } from "../MyContext";
 import { mutationResolvers } from "./Mutation";
 import { queryResolvers } from "./Query";
+import { reviewResolvers } from "./reviewResolvers";
 import { scalarResolvers } from "./scalars";
-import { subscriptionResolvers } from "./Subscription";
 
-export const resolvers: ResolverMap<MyContext> = {
-  ...queryResolvers,
-  ...mutationResolvers,
-  ...scalarResolvers,
-  ...subscriptionResolvers,
-};
+export const resolvers: ResolverMap<MyContext>[] = [
+  queryResolvers,
+  mutationResolvers,
+  scalarResolvers,
+  reviewResolvers,
+];