+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Add `GraphQLServer` class, combining HTTP and web socket server setup
 - Add `DirectivesMap` (maps directive names to their arguments) to code generation
 - Add simplified request handlers for Express and Koa
+- Allow lists of resolvers in `buildExecutableSchema`, merged via new utility `mergeResolvers` function
 
 ## v0.3.5